use asan and ubsan
This will be of particular value soon, since I'll be trying to allocate and use
the correct amount of space in the extents array for how many extents there
are.

I plan to avoid undefined behavior on race conditions, of course. But the extra
layer of checking is likely to help, at least with debugging.