added FIXME comment in Makefile about covering all clang versions
Finding a string in a makefile variable is easy. Finding it at just the
beginning of the match, or finding _where_ it matches, is harder, and I don't
know how to do it.

I know a way that works when sh is bash, but sh is not, in general, bash. In
particular, sh isn't bash on most Debian and Ubuntu systems, including my
development and testing system.

If I understood the quoting rules when running shell commands from a makefile,
I could do it, though I'm hoping there's a nicer way than that.