make nonzero size with no extents an error
It seems arbitrary to have other weird situations be errors but not this one.

Also, while the main use of this utility is to manually inspect the output, I
want it to be scriptable. So it's handy for harder-to-process outputs that
would typically prevent a calling script from succeeding to terminate with an
error, with nonzero exit status.