this is the idea for options (completely untested)
Currently, the table columns are the only thing the user can specify.

In theory, -t is implemented.

More options that abbreviate -t followed by common combinations of layout specifiers should be added.

The -h option for help should show an actual help message.

The help message should, at minimum, describe the behavior of the -t option and the (currently eight) characters that can go in a column specification string.

Long options should possibly be supported. But see the TODO in get_table_configuration() about that.

Before any of that, this should should of course be tested, and obvious bugs fixed.