implement the -V / --version option; replace some macros with consts
I know this is C, and the compiler doesn't treat consts with compile-time
initializers as compile-time constants. I still prefer the compiler be able to
reason about what's going on. More important, the specific code where I did
this is not a performance bottleneck and is very unlikely ever to become so.

This is a revision of an earlier pushed commit because the previous commit
message was extremely inadequate. It didn't mention -V at all. Sorry!