added FIXME about range of shell arithmetic
I believe integers for shell arithmetic in bash are represented byy the long
type in C. On GNU/Linux on x86 processors, long has the architcture's word
size. This is 8 bytes on x86_64 but only 4 bytes on x86_32.

This is supposed to work on various other architectures too, including 32-bit
armhf! This could be a problem.

For now, I want to fix bugs and get it working on *some* architecture (x86_64).
But the shell arithmetic range issue is a serious potential problem. A lot of
the numbers being represented with shell arithemtic have type __u64 from the
Linux kernel headers. On an x86_64 system, that's unsigned long long, which on
GNU/Linux has the same size and range as unsigned long. Many other
architectures don't have 64-bit longs.