avoid unintentional seeking on fd 3
I intend this as a fix for the problem investigated in the test-debug-cmp branch.

I suspect it might also fix the problem on my 16.04 LTS machine investigated in the test-debug-skip branch.

I'm not sure if it will fix either.

I want this saved and backed up before I test because it substantially changes the raw block device access behavior of a little-tested script that runs as root. :)