dedicated to public domain
In the event the software can't be public domain, it is license as permissively as practical.

This is achieved by adding CC0 notices to all files and including a ful copy of CC0 as COPYING.

I have also used this opportunity to apply and, occasionally, update descriptive comments at the tops of files.